refactor(frontend): split router config out of router.tsx
router.tsx mixed component definitions (RootShell, NavLink, NotFound)
with non-component exports (routeTree, createAppRouter, router), which
breaks Fast Refresh and triggered react-refresh/only-export-components
on main. Keep router.tsx component-only and move the route assembly to
a sibling router-config.ts; update main.tsx and App.test.tsx importers.
